requirements.txt
twscrape
asyncio
psycopg
google-generativeai
schedule
backoff
black
loguru
cryptography
colorama
python-dotenv
openai
aiosmtplib
setuptools
tavily-python
spacy
python-dateutil
psycopg[binary,pool]